Use `setup_test_component_platform` helper for date entity component tests instead of `hass.components`
