+++ tests/components/date/test_init.py
@@ -2,7 +2,7 @@
 
 from datetime import date
 
-from homeassistant.components.date import DOMAIN, SERVICE_SET_VALUE, DateEntity
+from homeassistant.components.date import DOMAIN, SERVICE_SET_VALUE
 from homeassistant.const import (
     ATTR_DATE,
     ATTR_ENTITY_ID,
@@ -12,25 +12,18 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.setup import async_setup_component
 
-
-class MockDateEntity(DateEntity):
-    """Mock date device to use in tests."""
-
-    _attr_name = "date"
-
-    def __init__(self, native_value=date(2020, 1, 1)) -> None:
-        """Initialize mock date entity."""
-        self._attr_native_value = native_value
-
-    async def async_set_value(self, value: date) -> None:
-        """Set the value of the date."""
-        self._attr_native_value = value
+from tests.common import setup_test_component_platform
+from tests.components.date.common import MockDateEntity
 
 
 async def test_date(hass: HomeAssistant, enable_custom_integrations: None) -> None:
     """Test date entity."""
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
+    entity = MockDateEntity(
+        name="test",
+        unique_id="unique_date",
+        native_value=date(2020, 1, 1),
+    )
+    setup_test_component_platform(hass, DOMAIN, [entity])
 
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
